pyproject.toml
[tool.poetry]
name = "mcp-embedded-docs"
version = "0.1.0"
description = "MCP server for embedded systems documentation"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "mcp_embedded_docs"}]

[tool.poetry.dependencies]
python = "^3.10"
pymupdf = "^1.23.0"
pdfplumber = "^0.10.0"
sentence-transformers = "^2.2.0"
faiss-cpu = "^1.7.4"
mcp = "^1.0.0"
pydantic = "^2.5.0"
openai = "^1.6.0"
httpx = "^0.27.0"
pyyaml = "^6.0"
click = "^8.1.0"
numpy = ">=1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.12.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mcp-embedded-docs = "mcp_embedded_docs.__main__:cli"